LIBFCREPO-116. Provision with development Loris.
The development version of Loris is currently on track to become the 2.1.0 release. This release contains, among other things, the ability to pull in configuration from environment variables, so the entire loris2.conf file is now under version control in iiif-env with just select portions left as environment variables to fill in.

https://issues.umd.edu/browse/LIBFCREPO-116